[rocRoller] Allow non power of two tile sizes (#730)
There is currently an issue with division when calculating the
`tileCeilDivide`. It sets the denominator to be an unsigned 32bit
integer. But the `sdSize` expression is a signed 64bit integer. This
causes problems with magic division. It also probably would lead to
errors if `sdSize` is larger than the max value a 32-bit integer can
hold.

This PR switches the denominator to be a signed 64bit integer as well. A
new powerOfTwoDivision function was needed for signed 64bit integers,
since one didn't already exist.

Added some GEMM tests that use non-power of 2 tile sizes.

+++ shared/rocroller/lib/source/KernelGraph/Utils.cpp
@@ -1137,7 +1137,7 @@ namespace rocRoller
 
         Expression::ExpressionPtr tileCeilDivide(Expression::ExpressionPtr sdSize, int tileSize)
         {
-            auto tileSizeExpr = Expression::literal(static_cast<uint>(tileSize));
+            auto tileSizeExpr = Expression::literal(static_cast<int64_t>(tileSize));
             auto one          = Expression::literal(1u);
 
             return (sdSize + tileSizeExpr - one) / tileSizeExpr;
